Changed volume type

+++ src/main/java/com/ldtsfeup2526/bobTheDestructor/sounds/VolumeControl.java
@@ -1,6 +1,6 @@
 package com.ldtsfeup2526.bobTheDestructor.sounds;
 
 public interface VolumeControl {
-    void setVolume(int volume);
-    int getVolume();
+    void setVolume(float volume);
+    float getVolume();
 }
